Remove domain namespace
