+++ app/interactions/force_delete_interaction/set_force_delete.rb
@@ -1,0 +1,12 @@
+module ForceDeleteInteraction
+  class SetForceDelete < Base
+    def execute
+      compose(CheckDiscarded, inputs)
+      compose(PrepareDomain, inputs)
+      compose(SetStatus, inputs)
+      compose(PostSetProcess, inputs)
+      compose(NotifyRegistrar, inputs)
+      compose(NotifyByEmail, inputs)
+    end
+  end
+end
